+++ src/Ext/TechnoType/Body.cpp
@@ -965,6 +965,10 @@ void TechnoTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 	this->DamagedSpeed.Read(exINI, pSection, "DamagedSpeed");
 	this->ProneSpeed.Read(exINI, pSection, "ProneSpeed");
 
+	this->CrushLevel.Read(exINI, pSection, "CrushLevel.%s");
+	this->CrushableLevel.Read(exINI, pSection, "CrushableLevel.%s");
+	this->DeployedCrushableLevel.Read(exINI, pSection, "DeployedCrushableLevel.%s");
+
 	this->SuppressKillWeapons.Read(exINI, pSection, "SuppressKillWeapons");
 	this->SuppressKillWeapons_Types.Read(exINI, pSection, "SuppressKillWeapons.Types");
 
@@ -1601,6 +1605,10 @@ void TechnoTypeExt::ExtData::Serialize(T& Stm)
 
 		.Process(this->DamagedSpeed)
 		.Process(this->ProneSpeed)
+
+		.Process(this->CrushLevel)
+		.Process(this->CrushableLevel)
+		.Process(this->DeployedCrushableLevel)
 
 		.Process(this->SuppressKillWeapons)
 		.Process(this->SuppressKillWeapons_Types)