+++ src/Ext/TechnoType/Body.h
@@ -354,6 +354,12 @@ public:
 		Nullable<double> ProneSpeed;
 		Nullable<double> DamagedSpeed;
 
+		Nullable<AnimTypeClass*> Promote_VeteranAnimation;
+		Nullable<AnimTypeClass*> Promote_EliteAnimation;
+    
+		Promotable<int> CrushLevel;
+		Promotable<int> CrushableLevel;
+		Promotable<int> DeployedCrushableLevel;
 		ValueableVector<AnimTypeClass*> Promote_VeteranAnimation;
 		ValueableVector<AnimTypeClass*> Promote_EliteAnimation;
 
@@ -759,6 +765,10 @@ public:
 
 			, ProneSpeed { }
 			, DamagedSpeed { }
+
+			, CrushLevel { -1 }
+			, CrushableLevel { -1 }
+			, DeployedCrushableLevel { -1 }
 
 			, SuppressKillWeapons { false }
 			, SuppressKillWeapons_Types { }