+++ src/Ext/Unit/Hooks.Crushing.cpp
@@ -1,11 +1,121 @@
 #include <DriveLocomotionClass.h>
 #include <ShipLocomotionClass.h>
 #include <UnitClass.h>
+#include <OverlayTypeClass.h>
 
 #include <Ext/Techno/Body.h>
-#include <Ext/TechnoType/Body.h>
 #include <Utilities/Macro.h>
 #include <Utilities/TemplateDef.h>
+
+#define Hook_IsCrusher(addr, name, mode, size, reg, retn, retn2) \
+DEFINE_HOOK(addr, ##name##_IsCrusher##mode##, size) \
+{ \
+	GET(FootClass*, pThis, reg); \
+	return TechnoExt::GetCrushLevel(pThis) > 0 ? retn : retn2; \
+}
+
+Hook_IsCrusher(0x4B3504, DriveLocomotionClass_PassableCheck, , 0x8, ECX, 0x4B3516, 0x4B3525)
+Hook_IsCrusher(0x4B414F, DriveLocomotionClass_PassableCheck, 2, 0x8, ECX, 0x4B4161, 0x4B4179)
+Hook_IsCrusher(0x5B098E, MechLocomotionClass_ProcessMoving, , 0x8, ECX, 0x5B09A0, 0x5B09AF)
+Hook_IsCrusher(0x5B1034, MechLocomotionClass_ProcessMoving, 2, 0x8, ECX, 0x5B1054, 0x5B108A)
+Hook_IsCrusher(0x5B1418, MechLocomotionClass_ProcessMoving, 3, 0x8, ECX, 0x5B1438, 0x5B146E)
+Hook_IsCrusher(0x6A1044, ShipLocomotionClass_ProcessMoving, , 0x8, ECX, 0x6A1064, 0x6A10B9)
+Hook_IsCrusher(0x6A2B53, ShipLocomotionClass_PassableCheck, , 0x8, ECX, 0x6A2B65, 0x6A2B74)
+Hook_IsCrusher(0x6A377B, ShipLocomotionClass_PassableCheck, 2, 0x8, ECX, 0x6A378D, 0x6A37A5)
+Hook_IsCrusher(0x73AFEB, UnitClass_PerCellProcess, , 0x6, EBP, 0x73B002, 0x73B074)
+Hook_IsCrusher(0x73FC6C, UnitClass_IsCellOccupied, , 0x6, EBX, 0x73FD37, 0x73FC91)
+Hook_IsCrusher(0x741733, UnitClass_CrushCell, , 0x6, EDI, 0x741754, 0x74195E)
+
+#undef Hook_IsCrusher
+
+DEFINE_JUMP(LJMP, 0x73FB2A, 0x73FB47)// Skip Crusher check before call ObjectClass::IsCrushable()
+
+DEFINE_JUMP(LJMP, 0x73FE5F, 0x73FE7C)// Skip Crusher check before call ObjectClass::IsCrushable()
+DEFINE_JUMP(LJMP, 0x741529, 0x741546)// Skip Crusher check before call ObjectClass::IsCrushable()
+
+DEFINE_HOOK(0x741603, UnitClass_ApproachTarget_OmniCrusher, 0x6)
+{
+	enum { IsOmniCrusher = 0x741613, NotOmniCrusher = 0x741685 };
+
+	GET(UnitClass*, pThis, ESI);
+
+	return TechnoExt::GetCrushLevel(pThis) >= RulesExt::Global()->OmniCrusherLevel ? IsOmniCrusher : NotOmniCrusher;
+}
+
+DEFINE_JUMP(LJMP, 0x7438F7, 0x743918)// Skip Crusher check before call ObjectClass::IsCrushable()
+
+DEFINE_HOOK(0x73B013, UnitClass_PerCellProcess_CrusherWall, 0x6)
+{
+	enum { CanCrush = 0x73B036, CannotCrush = 0x73B074 };
+
+	GET(OverlayTypeClass*, pOverlay, ESI);
+
+	// WW's code, ArrayIndex=0 is GASAND in default case, I guess it means sand bag is crushable
+	if (!pOverlay->ArrayIndex || pOverlay->Crushable)
+		return CanCrush;
+
+	if (!pOverlay->Wall)
+		return CannotCrush;
+
+	GET(UnitClass*, pThis, EBP);
+
+	return pThis->Type->MovementZone == MovementZone::CrusherAll || TechnoExt::GetCrushLevel(pThis) > RulesExt::Global()->WallCrushableLevel ? CanCrush : CannotCrush;
+}
+
+DEFINE_HOOK(0x73F42E, UnitClass_IsCellOccupied_CrushWall, 0x6)
+{
+	enum { CanCrush = 0x73F46E, CannotCrush = 0x73F483 };
+
+	GET(UnitClass*, pThis, EBX);
+
+	return pThis->Type->MovementZone == MovementZone::CrusherAll || TechnoExt::GetCrushLevel(pThis) > RulesExt::Global()->WallCrushableLevel ? CanCrush : CannotCrush;
+}
+
+DEFINE_HOOK(0x4B19B8, DriveLocomotionClass_ProcessMoving_CrushWall, 0x8)
+{
+	enum { CanCrush = 0x4B19E2, CannotCrush = 0x4B1A04 };
+
+	GET(FootClass*, pLinkedTo, ECX);
+	GET(OverlayTypeClass*, pOverlay, ESI);
+
+	return pOverlay->Wall && TechnoExt::GetCrushLevel(pLinkedTo) > RulesExt::Global()->WallCrushableLevel ? CanCrush : CannotCrush;
+}
+
+DEFINE_HOOK(0x4B1A1B, DriveLocomotionClass_ProcessMoving_CrusherAll, 0x8)
+{
+	enum { CanCrush = 0x4B1A2C, CannotCrush = 0x4B1A77 };
+
+	GET(FootClass*, pLinkedTo, ECX);
+
+	return pLinkedTo->GetTechnoType()->MovementZone == MovementZone::CrusherAll || TechnoExt::GetCrushLevel(pLinkedTo) > RulesExt::Global()->OmniCrusherLevel ? CanCrush : CannotCrush;
+}
+
+DEFINE_HOOK(0x5F6CD0, ObjectClass_IsCrushable, 0x6)
+{
+	enum { SkipGameCode = 0x5F6D90 };
+
+	GET(ObjectClass*, pThis, ECX);
+	GET_STACK(FootClass*, pCrusher, 0x4);
+	bool result = false;
+
+	if (pThis && pCrusher && pThis != pCrusher)
+	{
+		if (pThis->AbstractFlags & AbstractFlags::Techno)
+		{
+			const auto pFoot = abstract_cast<FootClass*>(pThis);
+
+			if (pFoot && !pCrusher->Owner->IsAlliedWith(pFoot) && !pFoot->IsIronCurtained())
+				result = TechnoExt::GetCrushLevel(pCrusher) > TechnoExt::GetCrushableLevel(pFoot);
+		}
+		else
+		{
+			result = pThis->GetType()->Crushable;
+		}
+	}
+
+	R->AL(result);
+	return SkipGameCode;
+}
 
 DEFINE_HOOK(0x73B05B, UnitClass_PerCellProcess_TiltWhenCrushes, 0x6)
 {